[bw] uint/ulong fix in binary code [bw] removed debug tempfile in lingen
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@623 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/bw/cxx/binary_ulong_traits.hpp
@@ -133,7 +133,7 @@ struct binary_pod_traits {
         static std::istream& get(std::istream& o, scalar_t & x)
         {
             T z = 0;
-            unsigned int v;
+            unsigned long v;
             for(unsigned int i = 0 ; i < nbits ; i++) {
                 o >> v;
                 z |= (v << i);
